RGBW PWM scalling in range to 1023

+++ src/supla/control/rgb_leds.h
@@ -25,11 +25,11 @@ class RGBLeds : public RGBBase {
  public:
   RGBLeds(int redPin, int greenPin, int bluePin);
 
-  void setRGBWValueOnDevice(uint8_t red,
-                            uint8_t green,
-                            uint8_t blue,
-                            uint8_t colorBrightness,
-                            uint8_t brightness);
+  void setRGBWValueOnDevice(uint32_t red,
+                            uint32_t green,
+                            uint32_t blue,
+                            uint32_t colorBrightness,
+                            uint32_t brightness);
 
   void onInit();
